Refactor sparkctl
