+++ cmd/sparkctl/create/s3.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package app
+package create
 
 import (
 	"context"
