+++ cmd/sparkctl/create/testdata/hadoop-conf/binary.dat
