+++ cmd/sparkctl/create/testdata/hadoop-conf/core-site.xml
